+++ Editor/IssueCategory.cs
@@ -1,0 +1,20 @@
+﻿namespace Unity.ProjectAuditor.Editor
+{
+    public enum IssueCategory
+    {
+        MetaData,
+        Assets,
+        Shaders,
+        ShaderVariants,
+        Code,
+        CodeCompilerMessages,
+        Generics,
+        ProjectSettings,
+        BuildFiles,
+        BuildSteps,
+        Assemblies,
+        Scenes,
+
+        NumCategories
+    }
+}