+++ Editor/UI/AreaSelectionWindow.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Unity.ProjectAuditor.Editor.UI.Framework;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
@@ -82,7 +83,7 @@ namespace Unity.ProjectAuditor.Editor.UI
             m_MultiColumnHeaderState = MultiSelectionTable.CreateDefaultMultiColumnHeaderState(headerData);
 
             var multiColumnHeader = new MultiColumnHeader(m_MultiColumnHeaderState);
-            multiColumnHeader.SetSorting((int)MultiSelectionTable.MyColumns.ItemName, true);
+            multiColumnHeader.SetSorting((int)MultiSelectionTable.Column.ItemName, true);
             multiColumnHeader.ResizeToFit();
             m_AreaTable = new MultiSelectionTable(m_TreeViewState, multiColumnHeader, names, selection);
         }