+++ Editor/UI/AssemblySelectionWindow.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Unity.ProjectAuditor.Editor.UI.Framework;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
@@ -85,7 +86,7 @@ namespace Unity.ProjectAuditor.Editor.UI
             m_MultiColumnHeaderState = MultiSelectionTable.CreateDefaultMultiColumnHeaderState(headerData);
 
             var multiColumnHeader = new MultiColumnHeader(m_MultiColumnHeaderState);
-            multiColumnHeader.SetSorting((int)MultiSelectionTable.MyColumns.ItemName, true);
+            multiColumnHeader.SetSorting((int)MultiSelectionTable.Column.ItemName, true);
             multiColumnHeader.ResizeToFit();
             m_MultiSelectionTable = new MultiSelectionTable(m_TreeViewState, multiColumnHeader, m_Names, selection);
         }