+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.6.2-preview] - 2021-05-25
+* Added Assemblies view (experimental)
 * Added Build Report Steps view
+* Added overview stats to Build Report Size view
 * Fixed detection of HDRP mixed LitShaderMode
 
 ## [0.6.1-preview] - 2021-05-11